+++ Test/SeablastMysqliTest.php
@@ -11,6 +11,13 @@ use Seablast\Seablast\SeablastMysqli;
 use Seablast\Seablast\Exceptions\DbmsException;
 use Tracy\Debugger;
 
+/**
+ * Basic tests of MySQLi connection.
+ * 
+ * Note: since PHPUnit 10 the method `expectWarning()` is removed without direct replacement,
+ * so there's no straightforward way to test the connection error behavior. I.e. failure of
+ * `$this->mysqli = new SeablastMysqli('invalid_host', 'user', 'password', 'database');`
+ */
 class SeablastMysqliTest extends TestCase
 {
     /** @var SeablastMysqli */
@@ -31,14 +38,6 @@ class SeablastMysqliTest extends TestCase
         $this->assertEquals('views', $configuration->getString(SeablastConstant::LATTE_TEMPLATE));
         $configuration->setInt(SeablastConstant::SB_LOGGING_LEVEL, 5);
         $configuration->setString(SeablastConstant::SB_PHINX_ENVIRONMENT, 'testing'); // so that the database test works
-        // Create local phinx configuration, so that it works on GitHub, yet can be adapted locally
-//        if (!file_exists('./conf/phinx.local.php') && file_exists('./conf/phinx.dist.php')) {
-//            if (copy('./conf/phinx.dist.php', './conf/phinx.local.php')) {
-//                Debugger::log('./conf/phinx.dist.php copied to ./conf/phinx.local.php', ILogger::INFO);
-//            } else {
-//                Debugger::log('./conf/phinx.dist.php failed to be copied to ./conf/phinx.local.php', ILogger::ERROR);
-//            }
-//        }
 
         $this->mysqli = $configuration->dbms();
     }
@@ -46,13 +45,6 @@ class SeablastMysqliTest extends TestCase
     public function testConstructorSuccess(): void
     {
         $this->assertInstanceOf(SeablastMysqli::class, $this->mysqli);
-    }
-
-    public function testConstructorThrowsExceptionOnConnectError(): void
-    {
-        $this->expectWarning();
-        // Note: Expecting E_WARNING and E_USER_WARNING is deprecated and will no longer be possible in PHPUnit 10.
-        $this->mysqli = new SeablastMysqli('invalid_host', 'user', 'password', 'database');
     }
 
     public function testQueryLogging(): void